+++ app/scripts/HiGlassComponent.js
@@ -9,7 +9,7 @@ import { ResizeSensor, ElementQueries } from 'css-element-queries';
 import * as PIXI from 'pixi.js';
 import vkbeautify from 'vkbeautify';
 import parse from 'url-parse';
-import createPubSub from 'pub-sub-es';
+import createPubSub, { globalPubSub } from 'pub-sub-es';
 
 import TiledPlot from './TiledPlot';
 import GenomePositionSearchBox from './GenomePositionSearchBox';
@@ -280,6 +280,7 @@ class HiGlassComponent extends React.Component {
     this.zoomHandlerBound = this.zoomHandler.bind(this);
     this.trackDroppedHandlerBound = this.trackDroppedHandler.bind(this);
     this.animateBound = this.animate.bind(this);
+    this.animateOnGlobalEventBound = this.animateOnGlobalEvent.bind(this);
     this.requestReceivedHandlerBound = this.requestReceivedHandler.bind(this);
     this.wheelHandlerBound = this.wheelHandler.bind(this);
     this.mouseMoveHandlerBound = this.mouseMoveHandler.bind(this);
@@ -679,6 +680,10 @@ class HiGlassComponent extends React.Component {
 
       this.isRequestingAnimationFrame = false;
     });
+  }
+
+  animateOnGlobalEvent({ sourceUid } = {}) {
+    if (sourceUid !== this.uid) this.animate();
   }
 
   measureSize() {
@@ -3296,15 +3301,27 @@ class HiGlassComponent extends React.Component {
       relTrackY: (hoveredTrack && hoveredTrack.flipText) ? relTrackPos[0] : relTrackPos[1],
       dataX,
       dataY,
-      isFrom2dTrack: hoveredTrack && hoveredTrack.is2d,
-      isFromVerticalTrack: hoveredTrack && hoveredTrack.flipText,
+      isFrom2dTrack: !!(hoveredTrack && hoveredTrack.is2d),
+      isFromVerticalTrack: !!(hoveredTrack && hoveredTrack.flipText),
       track: hoveredTrack,
       origEvt: e,
       sourceUid: this.uid,
       hoveredTracks,
+      noHoveredTracks: hoveredTracks.length === 0,
     };
 
     this.pubSub.publish('app.mouseMove', evt);
+
+    if (this.isBroadcastMousePositionGlobally) {
+      const eventDataOnly = { ...evt };
+      eventDataOnly.origEvt = undefined;
+      eventDataOnly.track = undefined;
+      eventDataOnly.hoveredTracks = undefined;
+      delete eventDataOnly.origEvt;
+      delete eventDataOnly.track;
+      delete eventDataOnly.hoveredTracks;
+      globalPubSub.publish('higlass.mouseMove', eventDataOnly);
+    }
 
     this.apiPublish('cursorLocation', {
       absX,